extracted some common logic
